app/requirements.txt
a2wsgi==1.10.8
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
altair==5.5.0
anyio==4.9.0
apache-airflow==2.5.0
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==3.3.2
argcomplete==3.6.1
asgiref==3.8.1
astor==0.8.1
attrs==25.3.0
babel==2.17.0
bleach==1.5.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==3.2.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
email-validator==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
html5lib==0.9999999
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
keras-layer-normalization==0.16.0
lazy-object-proxy==1.10.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgspec==0.19.0
multidict==6.2.0
namex==0.0.8
narwhals==1.32.0
numpy==2.2.4
optree==0.14.1
packaging==24.2
pandas==2.2.3
pathspec==0.9.0
pendulum==3.0.0
pillow==11.1.0
pluggy==1.5.0
prison==0.2.1
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.43.2
swagger_ui_bundle==1.1.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==1.8.0
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl#sha256=373c637f9c4f06346b1b02e280b30517d262a1651b252ff889f7c22716ef4548
tensorflow_keras==0.1
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
